Conflicts:
	kernel/kernel.c
	libc/include/sys/syscall.h

+++ libc/ioctl.c
@@ -1,8 +1,10 @@
 #include <sys/ioctl.h>
-#include <syscall.h>
+#include <sys/syscall.h>
+#include <sys/types.h>
+#include <unistd.h>
 
 int ioctl(int d, unsigned int request, void* data)
 {
-	syscall(SYS_IOCTL, d, request, data);
+	syscall(SYS_IOCTL, d, request, (uint32_t) data);
 	return 0;
 }
